+++ src/Rocks.Helpers/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.6.0 - 2019-11-12
+### Added
+- Added RetryOnExceptionAsync overload with CancellationToken 
+
 ## 4.5.0 - 2019-11-03
 ### Added
 - Added MergeIntoAsync